Split up SetOperation
